Move Files around, add Sentry and New relic
