+++ ManageATenancyAPI/Extension/ServiceConfiguration.cs
@@ -19,7 +19,9 @@ using ManageATenancyAPI.Services.Interfaces;
 using ManageATenancyAPI.Services.JWT;
 using ManageATenancyAPI.Tests;
 using ManageATenancyAPI.Tests.Unit.Services;
+using ManageATenancyAPI.UseCases.Meeting.GetMeeting;
 using ManageATenancyAPI.UseCases.Meeting.SaveMeeting;
+using ManageATenancyAPI.UseCases.Meeting.SignOffMeeting;
 
 namespace ManageATenancyAPI.Extension
 {
@@ -79,6 +81,10 @@ namespace ManageATenancyAPI.Extension
             services.AddScoped<ISaveEtraMeetingIssueGateway, SaveEtraMeetingIssueGateway>();
             services.AddScoped<ISaveEtraMeetingAttendanceGateway, SaveEtraMeetingAttendanceGateway>();
             services.AddScoped<ISaveEtraMeetingSignOffMeetingGateway, SaveEtraMeetingSignOffMeetingGateway>();
+
+            services.AddScoped<IGetEtraMeetingUseCase, GetEtraMeetingUseCase>();
+
+            services.AddScoped<ISignOffMeetingUseCase, SignOffMeetingUseCase>();
 
             services.AddScoped<IJpegPersistenceService, JpegPersistenceService>();
