+++ ManageATenancyAPI/ManageATenancyAPI.csproj
@@ -19,7 +19,6 @@
     <Folder Include="Plugins\" />
     <Folder Include="Services\S3\" />
     <Folder Include="Tests\Controllers\" />
-    <Folder Include="UseCases\Meeting\GetMeeting\Boundary\" />
     <Folder Include="wwwroot\" />
   </ItemGroup>
 
@@ -33,6 +32,8 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
     <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.0.1916590" />
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.1" />
+    <PackageReference Include="Sentry.AspNetCore" Version="1.1.2" />
+    <PackageReference Include="NewRelic.Agent" Version="8.13.798" />
     <PackageReference Include="Moq" Version="4.10.0" />
     <PackageReference Include="NLog.Web.AspNetCore" Version="4.7.0" />
     <PackageReference Include="Notify" Version="2.5.1" />
